requirements.txt
torchtyping
typeguard
transformers
gdown 
tqdm
timm
git+https://github.com/openai/CLIP.git
deepspeed
wandb
accelerate
galai
selenium
seleniumrequests
evaluate